use embedded tar instead of system-dependent tar for terminfo decompression to further achieve a self-contained executable

+++ src/ccdb/ccdb.set.cpp
@@ -215,7 +215,7 @@ INSTANTIATE_SET_PORT(mixedport,     "mixed-port")
 void ccdb::ccdb::set_log_size(const std::vector<std::string> &command_vector)
 {
     try {
-        max_log_size = std::strtol(command_vector.at(2).c_str(), nullptr, 10);
+        max_log_size = static_cast<int>(std::strtol(command_vector.at(2).c_str(), nullptr, 10));
     } catch (std::exception & e) {
         std::cerr << e.what() << std::endl;
     }
